Moved Spinner to separate file
Moved Spinner to separate file so it can be used by anyone.

To use the “Loading” indicator, simply add: <% include ./spinner %> to
your ejs file directly below the <% include ./header %> tag.

To show or hide the spinner, update $scope.showSpinner to true or false
in your controller or services file.